# app/docs/swagger.yaml
consumes:
- application/json
definitions:
  apperrors.ErrorCode:
    enum:
    - access_token_expired
    - authentication_error
    - authorization_error
    - database_error
    - forbidden
    - internal_error
    - invalid_request
    - malformed_body
    - not_implemented
    - password_too_short
    - refresh_token_invalid
    - resource_already_exists
    - resource_not_found
    - token_invalid
    type: string
    x-enum-varnames:
    - ErrCodeAccessTokenExpired
    - ErrCodeAuthenticationFailure
    - ErrCodeAuthorizationFailure
    - ErrCodeDatabaseError
    - ErrCodeForbidden
    - ErrCodeInternalError
    - ErrCodeInvalidRequest
    - ErrCodeMalformedBody
    - ErrCodeNotImplemented
    - ErrCodePasswordTooShort
    - ErrCodeRefreshTokenInvalid
    - ErrCodeResourceAlreadyExists
    - ErrCodeResourceNotFound
    - ErrCodeTokenInvalid
  auth.AccessTokenResponse:
    properties:
      access_token:
        example: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJpc3MiOiJTaWduYWxTZXJ2ZXIiLCJzdWIiOiI2OGZiNWY1Yi1lM2Y1LTRhOTYtOGQzNS1jZDIyMDNhMDZmNzMiLCJleHAiOjE3NDY3NzA2MzQsImlhdCI6MTc0Njc2NzAzNH0.3OdnUNgrvt1Zxs9AlLeaC9DVT6Xwc6uGvFQHb6nDfZs
        type: string
      account_id:
        example: a38c99ed-c75c-4a4a-a901-c9485cf93cf3
        type: string
      account_type:
        enum:
        - user
        - service_identity
        type: string
      expires_in:
        description: seconds
        example: 1800
        type: integer
      isn_perms:
        additionalProperties:
          type: string
        type: object
      role:
        enum:
        - owner
        - admin
        - member
        example: admin
        type: string
      token_type:
        example: Bearer
        type: string
    type: object
...
  handlers.CreateSignalsBatchResponse:
    properties:
      account_id:
        example: a38c99ed-c75c-4a4a-a901-c9485cf93cf3
        type: string
      resource_url:
        example: http://localhost:8080/api/isn/sample-isn--example-org/account/{account_id}/batch/{signals_batch_id}
        type: string
      signals_batch_id:
        example: b51faf05-aaed-4250-b334-2258ccdf1ff2
        type: string
    type: object
...
  utils.ErrorResponse:
    properties:
      error_code:
        allOf:
        - $ref: '#/definitions/apperrors.ErrorCode'
        example: example_error_code
      message:
        type: string
    type: object
host: localhost:8080
info:
  contact: {}
  description: Signals ISN service API
paths:
...
  /api/isn/{isn_slug}/signals/batches:
    post:
      description: |
        This endpoint is used by service accounts to create a new batch used to track signals sent to the specified isn

        For user accounts, a batch is automatically created when they are granted write permission to an isn and is only closed if their permission to write to the isn is revoked

        For service accounts, the client app can decide how long to keep a batch open
        (a batch status summary is sent to a webhook after the batch closes)

        opening a batch closes the previous batch created on the isn for this account.

        Signals can only be sent to open batches.

        authentication is based on the supplied access token:
        (the site owner; the isn admin and members with an isn_perm= write can create a batch)
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/handlers.CreateSignalsBatchResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utils.ErrorResponse'
      security:
      - BearerAccessToken: []
      summary: Create a new signal batch
      tags:
      - Signals Management
...
  /auth/revoke:
    post:
      description: |
        Revoke a refresh token to prevent it being used to create new access tokens.

        You need to supply a vaild refresh token to use this API - if the refresh token has expired or been revoked the user must login again to get a new one.

        The refresh token should be supplied in a http-only cookie called refresh_token.

        You must also provide a previously issued bearer access token - it does not matter if it has expired
        (the token is not used to authenticate the request but is needed to establish the ID of the user making the request)

        Note that any unexpired access tokens issued for this user will continue to work until they expire.
        Users must log in again to obtain a new refresh token if the current one has been revoked.
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/utils.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utils.ErrorResponse'
      security:
      - BearerRefreshToken: []
      summary: Revoke refresh token
      tags:
      - auth
  /auth/token:
    post:
      description: |
        Use this endpoint to get a new access token.

        You need to supply a vaild refresh token to use this API - if the refresh token has expired or been revoked the user must login again to get a new one.

        The refresh token should be supplied in a http-only cookie called refresh_token.

        You must also provide a previously issued bearer access token - it does not matter if it has expired
        (the token is not used to authenticate the request but is needed to establish the ID of the user making the request)

        Note this action automatically revokes the current refresh_token and issues a new one.

        The new refresh token is sent in an http-only cookie named refresh_token.
        In production deployments the secure flag (https only) on the cookie will be set to true

        Access tokens expire after 30 mins and subsequent requests using the token will fail with an error_code of "access_token_expired"
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/auth.AccessTokenResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/utils.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utils.ErrorResponse'
      security:
      - BearerRefreshToken: []
      summary: Refresh access token
      tags:
      - auth
...
produces:
- application/json
securityDefinitions:
  BearerAccessToken:
    description: Bearer {JWT access token}
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"
...